Revert "Third person view"
