--- .gitignore
@@ -1,6 +1,0 @@
-.idea/workspace.xml
-.idea/vcs.xml
-.idea/modules.xml
-.idea/misc.xml
-.idea/Projekt-WebGL.iml
-.idea/.gitignore
