+++ Physics.js
@@ -13,7 +13,7 @@ export default class Physics {
                 vec3.scaleAndAdd(node.translation, node.translation, node.velocity, dt);
                 node.updateTransform();
                 this.scene.traverse(other => {
-                    if (node !== other && !node.ignoreCollision && !other.ignoreCollision) {
+                    if (node !== other) {
                         this.resolveCollision(node, other);
                     }
                 });
@@ -90,7 +90,6 @@ export default class Physics {
 
         vec3.add(a.translation, a.translation, minDirection);
         a.updateTransform();
-
     }
 
 }
