+++ SceneBuilder.js
@@ -3,7 +3,6 @@ import Mesh from './Mesh.js';
 import Node from './Node.js';
 import Model from './Model.js';
 import Camera from './Camera.js';
-import Player from './Player.js';
 
 import Scene from './Scene.js';
 
@@ -15,18 +14,11 @@ export default class SceneBuilder {
 
     createNode(spec) {
         switch (spec.type) {
-            //case 'camera': return new Camera(spec);
+            case 'camera': return new Camera(spec);
             case 'model': {
                 const mesh = new Mesh(this.spec.meshes[spec.mesh]);
                 const texture = this.spec.textures[spec.texture];
                 return new Model(mesh, texture, spec);
-            }
-            case 'player' : {
-                const mesh = new Mesh(this.spec.meshes[spec.mesh]);
-                const texture = this.spec.textures[spec.texture];
-                let player = new Player(mesh, texture, spec);
-                player.addChild(new Camera(spec.children[0]))
-                return player;
             }
             default: return new Node(spec);
         }
