+++ scene.json
@@ -1,29 +1,16 @@
 {
   "nodes": [
     {
-      "type": "player",
-      "mesh": 1,
-      "texture": 1,
-      "aabb": {
-        "min": [-0.5, -0.5, -0.5],
-        "max": [0.5, 0.5, 0.5]
-      },
+      "type": "camera",
       "translation": [0, 1, 0],
-      "scale": [1, 1, 1],
-      "children": [
-        {
-          "type": "camera",
-          "translation": [0, 0.4, 1.5],
-          "aabb": {
-            "min": [0, 0, 0],
-            "max": [0, 0, 0]
-          },
-          "aspect": 1,
-          "fov": 1.5,
-          "near": 0.01,
-          "far": 100
-        }
-      ]
+      "aabb": {
+        "min": [-0.2, -0.2, -0.2],
+        "max": [0.2, 0.2, 0.2]
+      },
+      "aspect": 1,
+      "fov": 1.5,
+      "near": 0.01,
+      "far": 100
     },
     {
       "type": "model",
